+++ CHANGELOG.md
@@ -1,7 +1,16 @@
+### 0.1.2
+
+- handling variants as an array of strings [#1]
+- add demo more examples
+- output terser
+- docs update
+
 ### 0.1.1
-* removed unnecessary dependencies
-* add the demo missing types
-* docs update
+
+- removed unnecessary dependencies
+- add the demo missing types
+- docs update
 
 ### 0.1.0
-* initial package state
+
+- initial package state